+++ as-pts1/travail/projet test/Classes/Header/Dagger.hpp
@@ -10,7 +10,7 @@
 class Dagger : public Weapon
 {
 public:
-	virtual bool attack(HittableCharacter* lanceur, const sf::Vector2i& attackPoint) override;
+	virtual bool attack(HittableCharacter* lanceur, const sf::Vector2f& attackPoint) override;
 
 };
