+++ as-pts1/travail/projet test/Classes/Resource/Arrow.cpp
@@ -14,7 +14,7 @@ Arrow::Arrow()
 
 }
 
-Arrow::Arrow(const sf::Vector2i& launchDirection, double damage, const sf::Vector2f& startPosition,float vitesse,float dureeVie) {
+Arrow::Arrow(const sf::Vector2f& launchDirection, double damage, const sf::Vector2f& startPosition,float vitesse,float dureeVie) {
 	_damage = damage;
 	speed = vitesse;
 	nextDirection = sf::Vector2f(launchDirection);
