+++ as-pts1/travail/projet test/Classes/Resource/Dagger.cpp
@@ -5,13 +5,13 @@
 #include "..\Header\Arrow.hpp"
 #include "..\Header\GameMaster.hpp"
 
-bool Dagger::attack(HittableCharacter* lanceur, const sf::Vector2i& attackPoint)
+bool Dagger::attack(HittableCharacter* lanceur, const sf::Vector2f& attackPoint)
 {
 
     //création du projectile et ajout de ce projectile dans la liste des objets du GameMaster
-    sf::Vector2i lanceurPos = sf::Vector2i(lanceur->getSprite().getPosition());
+    sf::Vector2f lanceurPos = sf::Vector2f(lanceur->getSprite().getPosition());
 
-    sf::Vector2i launchDirection = attackPoint - lanceurPos;
+    sf::Vector2f launchDirection = attackPoint - lanceurPos;
 
     auto myProjectile = new Arrow(launchDirection, 1, lanceur->getSprite().getPosition(), _attackSpeed, _attackDuration);
